(mon-pix)(wip): Create session service and use handleAuthentication method in routes
