+++ mon-pix/app/routes/campaigns/access.js
@@ -27,17 +27,8 @@ export default class AccessRoute extends Route {
     } else if (this._shouldJoinSimplifiedCampaignAsAnonymous(campaign)) {
       this.authenticationRoute = 'campaigns.join.anonymous';
     }
-    const isUserLoaded = !!this.currentUser.user;
-    const isAuthenticated = this.session.get('isAuthenticated');
-    if (!isAuthenticated || !isUserLoaded) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo(this.authenticationRoute);
-    } else if (this.currentUser.user.mustValidateTermsOfService) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('terms-of-service');
-    } else {
-      return super.beforeModel(...arguments);
-    }
+
+    this.session.requireAuthenticationAndApprovedTermsOfService(transition, this.authenticationRoute);
   }
 
   model() {
