+++ mon-pix/app/routes/campaigns/assessment/start-or-resume.js
@@ -9,19 +9,8 @@ export default class EvaluationStartOrResumeRoute extends Route {
   userHasJustConsultedTutorial = false;
 
   beforeModel(transition) {
-    const isUserLoaded = !!this.currentUser.user;
-    const isAuthenticated = this.session.get('isAuthenticated');
-    if (!isAuthenticated || !isUserLoaded) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('login');
-    } else if (this.currentUser.user.mustValidateTermsOfService) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('terms-of-service');
-    } else {
-      return super.beforeModel(...arguments);
-    }
     this.userHasJustConsultedTutorial = transition.to.queryParams.hasConsultedTutorial;
-    super.beforeModel(...arguments);
+    this.session.requireAuthenticationAndApprovedTermsOfService(transition);
   }
 
   async model() {
