+++ mon-pix/app/routes/campaigns/invited/fill-in-participant-external-id.js
@@ -5,20 +5,9 @@ export default class FillInParticipantExternalIdRoute extends Route {
   @service campaignStorage;
   @service session;
   @service router;
-  @service currentUser;
 
   beforeModel(transition) {
-    const isUserLoaded = !!this.currentUser.user;
-    const isAuthenticated = this.session.get('isAuthenticated');
-    if (!isAuthenticated || !isUserLoaded) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('login');
-    } else if (this.currentUser.user.mustValidateTermsOfService) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('terms-of-service');
-    } else {
-      return super.beforeModel(...arguments);
-    }
+    this.session.requireAuthenticationAndApprovedTermsOfService(transition);
   }
 
   async model() {
