+++ mon-pix/app/routes/campaigns/invited/student-sup.js
@@ -9,17 +9,7 @@ export default class StudentSupRoute extends Route {
   @service router;
 
   beforeModel(transition) {
-    const isUserLoaded = !!this.currentUser.user;
-    const isAuthenticated = this.session.get('isAuthenticated');
-    if (!isAuthenticated || !isUserLoaded) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('login');
-    } else if (this.currentUser.user.mustValidateTermsOfService) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('terms-of-service');
-    } else {
-      return super.beforeModel(...arguments);
-    }
+    this.session.requireAuthenticationAndApprovedTermsOfService(transition);
   }
 
   model() {
