+++ mon-pix/app/routes/certifications/join.js
@@ -5,20 +5,9 @@ import Route from '@ember/routing/route';
 export default class JoinRoute extends Route {
   @service currentUser;
   @service session;
-  @service router;
 
   beforeModel(transition) {
-    const isUserLoaded = !!this.currentUser.user;
-    const isAuthenticated = this.session.get('isAuthenticated');
-    if (!isAuthenticated || !isUserLoaded) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('login');
-    } else if (this.currentUser.user.mustValidateTermsOfService) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('terms-of-service');
-    } else {
-      return super.beforeModel(...arguments);
-    }
+    this.session.requireAuthenticationAndApprovedTermsOfService(transition);
   }
 
   model() {
