+++ mon-pix/app/routes/certifications/results.js
@@ -4,21 +4,9 @@ import { inject as service } from '@ember/service';
 export default class ResultsRoute extends Route {
   @service session;
   @service store;
-  @service router;
-  @service currentUser;
 
   beforeModel(transition) {
-    const isUserLoaded = !!this.currentUser.user;
-    const isAuthenticated = this.session.get('isAuthenticated');
-    if (!isAuthenticated || !isUserLoaded) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('login');
-    } else if (this.currentUser.user.mustValidateTermsOfService) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('terms-of-service');
-    } else {
-      return super.beforeModel(...arguments);
-    }
+    this.session.requireAuthenticationAndApprovedTermsOfService(transition);
   }
 
   async model(params) {
