+++ mon-pix/app/routes/user-tests.js
@@ -10,17 +10,7 @@ export default class UserTestsRoute extends Route {
   @service router;
 
   beforeModel(transition) {
-    const isUserLoaded = !!this.currentUser.user;
-    const isAuthenticated = this.session.get('isAuthenticated');
-    if (!isAuthenticated || !isUserLoaded) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('login');
-    } else if (this.currentUser.user.mustValidateTermsOfService) {
-      this.session.set('attemptedTransition', transition);
-      this.router.transitionTo('terms-of-service');
-    } else {
-      return super.beforeModel(...arguments);
-    }
+    this.session.requireAuthenticationAndApprovedTermsOfService(transition);
   }
 
   model() {
